Determine the density according to DisplayDevice (q-x86)
Previously we use hwConfig to get the dimension of the display.
However, there may be several hwConfig modes of a display. Only
the first mode is used to calculate the density. But first mode
may not be the default mode. Hence the calculated density may not
be suitable to the default mode.

Now use the dimension of the DisplayDevice to calculate the density.

(q-x86) Changelog:
        Adapt code for change from 'static int density' to 'static float density'
        Remove duplicated info.{w,h} assignements
        Keep info.{w,h} and info.viewport{W,H} assignements as they are
        Move up 'const auto display = getDefaultDisplayDeviceLocked();'
        Use 'display' variable instead of 'hw' in SurfaceFlinger::getDisplayConfigs